feat(ts): convert to typescript and prepare publish
