+++ README.md
@@ -1,0 +1,11 @@
+# bool-is-bool
+
+## Make bool boolable again! :hand: :gem: :hand: :rocket: :moon:
+
+Check if any value is a real bool-bool!
+
+All written in Typescript, ESFuture compliant!
+
+With lots of dev-deps!
+
+Shipped with 100% test-coverage!
