+++ package.json
@@ -2,17 +2,28 @@
   "name": "bool-is-bool",
   "version": "1.0.1",
   "description": "Check if a value that seems a bool, is really a bool :)",
-  "main": "src/index.js",
+  "readme": "https://github.com/finkinfridom/bool-is-bool/README.md",
+  "main": "dist/index.js",
+  "files": [
+    "/dist"
+  ],
   "repository": "https://github.com/finkinfridom/bool-is-bool",
   "author": "Alberto Maghini <a.maghini@gmail.com>",
   "license": "MIT",
   "scripts": {
-    "test": "mocha __tests__/**/*.js",
-    "watch": "DEBUG=bool-is-bool mocha __tests__/**/*.js --watch"
+    "test": "mocha -r ts-node/register ./__tests__/*.spec.ts",
+    "watch": "DEBUG=bool-is-bool mocha -r ts-node/register ./__tests__/*.spec.ts --watch",
+    "pack": "tsc"
   },
   "devDependencies": {
+    "@types/chai": "^4.3.3",
+    "@types/debug": "^4.1.7",
+    "@types/mocha": "^9.1.1",
+    "chai": "^4.3.6",
     "debug": "4.3.4",
     "mocha": "10.0.0",
-    "supports-color": "9.2.3"
+    "supports-color": "9.2.3",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.8.3"
   }
 }
